+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * config cleanup
   * removed deprecated @owner and @maintainer contact shortcuts
   * renamed `ldapSso` to `ldapUnit`
+  * removed "downloadsCache" property, use symlinks for $home/downloads instead
 * .backstage cleanup
   * renamed `maintainer` to `modified`
   * renamed `creator` field to `created-by` and `created` to `created-at`